requirements.txt
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0,!=1.12.0
torchvision>=0.8.1,!=0.13.0
tqdm>=4.41.0
protobuf<4.21.3
pandas>=1.1.4
ipython
psutil
func-timeout
flask
gunicorn
black
pydrive
requests
dash>=2.0.0
greenlet
click==6.7
itsdangerous==0.24
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
discord.py